app_b/requirements.txt
Flask
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-requests
opentelemetry-semantic-conventions


opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
